# inventory-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warehouse Inventory API
  description: RESTful API for managing warehouse inventory with MSSQL backend
  version: 1.0.0
  contact:
    name: OpenChoreo Demo
servers:
  - url: http://localhost:9090
    description: Local development server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API and database connection
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /inventory:
    get:
      summary: Get all inventory items
      description: Retrieves all products from the inventory
      operationId: getInventory
      tags:
        - Inventory
      responses:
        '200':
          description: List of all inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new inventory item
      description: Adds a new product to the inventory
      operationId: createInventoryItem
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /inventory/{id}:
    get:
      summary: Get inventory item by ID
      description: Retrieves a specific product by its ID
      operationId: getInventoryItem
      tags:
        - Inventory
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update inventory quantity
      description: Updates the stock quantity for a specific product
      operationId: updateInventoryQuantity
      tags:
        - Inventory
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuantityRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete inventory item
      description: Removes a product from the inventory
      operationId: deleteInventoryItem
      tags:
        - Inventory
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted successfully
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          description: Unique product identifier
          example: 1
        sku:
          type: string
          description: Stock keeping unit
          example: "WH-001-A"
        name:
          type: string
          description: Product name
          example: "Industrial Widget"
        quantity:
          type: integer
          description: Current stock quantity
          example: 150
        category:
          type: string
          description: Product category
          example: "Electronics"
        location:
          type: string
          description: Warehouse location
          example: "Aisle 3, Shelf B"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-01-22T10:30:00Z"

    CreateProductRequest:
      type: object
      required:
        - sku
        - name
      properties:
        sku:
          type: string
          description: Stock keeping unit
          example: "WH-001-A"
        name:
          type: string
          description: Product name
          example: "Industrial Widget"
        quantity:
          type: integer
          minimum: 0
          description: Initial stock quantity
          example: 100
        category:
          type: string
          description: Product category
          example: "Electronics"
        location:
          type: string
          description: Warehouse location
          example: "Aisle 3, Shelf B"

    UpdateQuantityRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 0
          description: New stock quantity
          example: 75

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          example: "2025-01-22T10:30:00Z"
        database:
          type: string
          example: "healthy"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: "not_found"
        message:
          type: string
          description: Error message
          example: "Product not found"